changed instances of uint to uint256; fixes issue #226

+++ test/helpers/MultisigWalletMock.sol
@@ -2,9 +2,9 @@ pragma solidity ^0.4.11;
 import "../../contracts/MultisigWallet.sol";
 
 contract MultisigWalletMock is MultisigWallet {
-  uint public totalSpending;
+  uint256 public totalSpending;
 
-  function MultisigWalletMock(address[] _owners, uint _required, uint _daylimit)
+  function MultisigWalletMock(address[] _owners, uint256 _required, uint256 _daylimit)
     MultisigWallet(_owners, _required, _daylimit) payable { }
 
   function changeOwner(address _from, address _to) external { }
