+++ test/helpers/SafeMathMock.sol
@@ -5,17 +5,17 @@ import '../../contracts/math/SafeMath.sol';
 
 
 contract SafeMathMock {
-  uint public result;
+  uint256 public result;
 
-  function multiply(uint a, uint b) {
+  function multiply(uint256 a, uint256 b) {
     result = SafeMath.mul(a, b);
   }
 
-  function subtract(uint a, uint b) {
+  function subtract(uint256 a, uint256 b) {
     result = SafeMath.sub(a, b);
   }
 
-  function add(uint a, uint b) {
+  function add(uint256 a, uint256 b) {
     result = SafeMath.add(a, b);
   }
 }
