+++ test/helpers/StandardTokenMock.sol
@@ -7,7 +7,7 @@ import '../../contracts/token/StandardToken.sol';
 // mock class using StandardToken
 contract StandardTokenMock is StandardToken {
 
-  function StandardTokenMock(address initialAccount, uint initialBalance) {
+  function StandardTokenMock(address initialAccount, uint256 initialBalance) {
     balances[initialAccount] = initialBalance;
     totalSupply = initialBalance;
   }
